Fix a few cross-platform issues
