+++ code/model/modelrender.h
@@ -11,7 +11,7 @@
 #define _MODELRENDER_H
 
 #include "model/model.h"
-#include "Math/vecmat.h"
+#include "math/vecmat.h"
 #include "lighting/lighting.h"
 #include "graphics/gropengltnl.h"
 
@@ -25,8 +25,30 @@ extern vec3d Object_position;
 
 extern team_color* Current_team_color;
 
-extern inline int in_sphere(vec3d *pos, float radius, vec3d *view_pos);
-extern inline int in_box(vec3d *min, vec3d *max, vec3d *pos, vec3d *view_pos);
+inline int in_box(vec3d *min, vec3d *max, vec3d *pos, vec3d *view_pos)
+{
+	vec3d point;
+
+	vm_vec_sub(&point, view_pos, pos);
+
+	if ( (point.xyz.x >= min->xyz.x) && (point.xyz.x <= max->xyz.x)
+		&& (point.xyz.y >= min->xyz.y) && (point.xyz.y <= max->xyz.y)
+		&& (point.xyz.z >= min->xyz.z) && (point.xyz.z <= max->xyz.z) )
+	{
+		return 1;
+	}
+
+	return -1;
+}
+
+inline int in_sphere(vec3d *pos, float radius, vec3d *view_pos)
+{
+	if ( vm_vec_dist(view_pos, pos) <= radius )
+		return 1;
+	else
+		return -1;
+}
+
 extern int model_interp_get_texture(texture_info *tinfo, fix base_frametime);
 
 struct transform
